cli: factor progress reporting out into its own module
